+++ CHANGES.rst
@@ -9,6 +9,8 @@ Changelog
 
 * [yarn] Set minimal requirement to 1.20.0.
 * [yarn] Fix deprecated global arguments.
+* [bitbar] Force refresh of local package databases before fetching outdated
+  ones.
 * [mpm] Add utilities to read a config TOML file. Refs #66.
 * [mpm] Auto-format Python code with Black.
 * [mpm] Move ``pytest`` config from setup.py to ``pyproject.toml``.